+++ packages/paste-codemods/transforms/__testfixtures__/barreled-to-unbarreled.input.ts
@@ -1,0 +1,6 @@
+import {Button, Label, MediaObject, TBodyPropTypes} from '@twilio-paste/core';
+import {Table} from '@twilio-paste/core';
+import {Button as Stuffin} from '@twilio-paste/core';
+import {Button as Stuff} from '@sendgrid/ui-components';
+import {Input} from '@twilio-paste/core';
+import {Card} from '@twilio-paste/core/dist/card';