+++ packages/paste-core/core-bundle/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.7](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@4.1.6...@twilio-paste/core@4.1.7) (2020-11-11)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [4.1.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@4.1.5...@twilio-paste/core@4.1.6) (2020-11-10)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [4.1.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@4.1.4...@twilio-paste/core@4.1.5) (2020-11-10)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
 ## [4.1.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@4.1.3...@twilio-paste/core@4.1.4) (2020-11-10)
 
 **Note:** Version bump only for package @twilio-paste/core